Refactor and image showed on dialog fragment
