+++ app/src/main/java/com/example/androidmobiletestapp/adapter/PlanetAdapter.kt
@@ -3,7 +3,7 @@ package com.example.androidmobiletestapp.adapter
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.androidmobiletestapp.PlanetResponse
+import com.example.androidmobiletestapp.data.PlanetResponse
 import com.example.androidmobiletestapp.R
 
 class PlanetAdapter(private val planetList:List<PlanetResponse>, private val onClickListener: (PlanetResponse) -> Unit): RecyclerView.Adapter<PlanetViewHolder>() {
