+++ app/src/main/java/com/example/androidmobiletestapp/api/APIService.kt
@@ -1,5 +1,6 @@
-package com.example.androidmobiletestapp
+package com.example.androidmobiletestapp.api
 
+import com.example.androidmobiletestapp.data.PlanetResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Url
